Aprimora cadastro de produtos e interface do usuário
- Altera tipos de `Quantidade` e `Preco` para `double` em `Produto.cs` para maior precisão.
- Atualiza título da página em `ListaProduto.xaml` para "Minhas Compras" e adiciona itens de toolbar.
- Implementa navegação para `NovoProduto` em `ListaProduto.xaml.cs`.
- Modifica título e adiciona campos de entrada em `NovoProduto.xaml`.
- Implementa lógica de salvamento de produto em `NovoProduto.xaml.cs` com tratamento de erros.

+++ MauiAppMinhasCompras/Models/Produto.cs
@@ -9,8 +9,8 @@ namespace MauiAppMinhasCompras.Models
 
         public string Descricao { get; set; }
 
-        public int Quantidade { get; set; }
+        public double Quantidade { get; set; }
 
-        public decimal Preco { get; set; }
+        public double Preco { get; set; }
     }
 }
